refactor: 이벤트 수정 페이지 - 참가자 목록 컴포넌트 분리
